backend/requirements.txt
flask==2.2.3
flask-socketio==5.3.3
flask-cors==3.0.10
psycopg2-binary==2.9.6
python-dotenv==1.0.0
langchain==0.0.267
pyjwt==2.6.0
werkzeug==2.2.3
gunicorn==20.1.0
eventlet==0.33.3
